requirements.txt
altgraph==0.17.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
constants==0.6.0
contourpy==1.3.1
cycler==0.11.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.55.3
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numexpr==2.11.0
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.1
pefile==2023.2.7
pickleshare==0.7.5
pillow==11.3.0
pip==25.2
pipreqs==0.5.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.7.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pyzmq==27.1.0
qtwidgets==1.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.15.3
setuptools==78.1.1
shiboken6==6.7.3
sip==6.10.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
yarg==0.1.9

flask
python-dotenv